+++ src/main/java/com/khozema/iniitian/dao/CategoryDAO.java
@@ -7,11 +7,11 @@ import com.khozema.iniitian.entity.Product;
 
 public interface CategoryDAO {
 
-	public Category get(Long id);
+	public Category get(int id);
 	public List<Category> getAll();
-	public List<Product> getProducts(Long id);
+	public List<Product> getProducts(int id);
 	public void add(Category c);
 	public void update(Category c);
-	public void remove(Long id);
+	public void remove(int id);
 	
 }
