+++ src/main/java/com/khozema/iniitian/service/ProductService.java
@@ -6,12 +6,12 @@ import com.khozema.iniitian.entity.Category;
 import com.khozema.iniitian.entity.Product;
 
 public interface ProductService {
-	public Product get(Long id);
+	public Product get(int id);
 	public List<Product> getAll();
-	public Category getCategory(Long id);
-	public List<Product> getProductsByCategory(Long id);
+	public Category getCategory(int id);
+	public List<Product> getProductsByCategory(int id);
 	public void add(Product c);
 	public void update(Product c);
-	public void remove(Long id);	
+	public void remove(int id);	
 
 }
